refactored the structure for components
