+++ practice-redux/src/Components/Link.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Link = ({ active, children, onClick }) => {
+    if (active) {
+        return <span>{children}</span>
+    }
+    return (
+        <a href="javascript:void(0)" onClick={(e) => {
+            e.preventDefault();
+            onClick();
+        }}> {children} </a>
+    );
+}
+
+export default Link;
